requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
uwsgi
flask-cors
python-dotenv
pymysql
jsonschema